refactor(slacksvc): move notify DM seam to notifier_notify.go (PRD #1026 M1)
